Added forecast function
forecastFunction(coordinates) function uses a different API endpoint than search(city) function. Requires lon&lat data of the searched city so in displayTemperature() call the forecastFunction(response.data.coord) to get the coordinates data.
The data is then used by the forecastFunction, this function then uses axios to get the ApiUrl, and once it has the response to displayForecast() function.